+++ R/PivotDataGroup.R
@@ -535,9 +535,9 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
      # check sort order
      if((dataSortOrder=="custom")&&(is.null(customSortOrder))) stop("PivotDataGroup$addDataGroups(): if dataSortOrder=custom, then customSortOrder must be specified.", call. = FALSE)
      # check outline settings
-     outlineBefore <- private$cleanOutlineArg(outlineBefore)
-     outlineAfter <- private$cleanOutlineArg(outlineAfter, defaultCaption="")
-     outlineTotal <- private$cleanOutlineArg(outlineTotal, defaultCaption="Total", defaultIsEmpty=FALSE)
+     outlineBefore <- cleanOutlineArg(private$p_parentPivot, outlineBefore)
+     outlineAfter <- cleanOutlineArg(private$p_parentPivot, outlineAfter, defaultCaption="")
+     outlineTotal <- cleanOutlineArg(private$p_parentPivot, outlineTotal, defaultCaption="Total", defaultIsEmpty=FALSE)
      # visual totals
      if(addTotal==TRUE){ private$p_visualTotals <- visualTotals }
      # if onlyAddGroupIf is specified, then preGroupData cannot be used
@@ -1063,8 +1063,8 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
        stop(paste0("PivotDataGroup$addCalculationGroups():  There are no visible calculations in the calculation group '", calculationGroupName, "'"), call. = FALSE)
      calculations <- calculations[order(names(calculations))]
      # outline info
-     outlineBefore <- private$cleanOutlineArg(outline=outlineBefore)
-     outlineAfter <- private$cleanOutlineArg(outline=outlineAfter, defaultCaption="")
+     outlineBefore <- cleanOutlineArg(private$p_parentPivot, outline=outlineBefore)
+     outlineAfter <- cleanOutlineArg(private$p_parentPivot, outline=outlineAfter, defaultCaption="")
      # where are the new groups being added?
      if(is.null(atLevel)) {
        # get the current set of leaf groups
@@ -1647,80 +1647,6 @@ PivotDataGroup <- R6::R6Class("PivotDataGroup",
    p_isRendered = FALSE, # helper flag to keep track of which data groups have already been rendered
 
    # private functions:
-   cleanOutlineArg = function(outline=NULL, defaultCaption="{value}", defaultIsEmpty=TRUE) { # checks values and provides defaults for settings (so following code does not need is.null checks)
-      if(private$p_parentPivot$argumentCheckMode > 0) {
-         checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotDataGroup", "cleanOutlineArg", outline, missing(outline), allowMissing=FALSE, allowNull=TRUE, allowedClasses=c("logical", "list"))
-         checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotDataGroup", "cleanOutlineArg", defaultCaption, missing(defaultCaption), allowMissing=TRUE, allowNull=FALSE, allowedClasses="character")
-         checkArgument(private$p_parentPivot$argumentCheckMode, FALSE, "PivotDataGroup", "cleanOutlineArg", defaultIsEmpty, missing(defaultIsEmpty), allowMissing=TRUE, allowNull=FALSE, allowedClasses="logical")
-      }
-      if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotDataGroup$cleanOutlineArg", "Cleaning outline argument...")
-      if("logical" %in% class(outline)) {
-         if(isTRUE(outline)) outline <- list()
-         else outline <- NULL
-      }
-      clean <- list()
-      # global switch
-      if(is.null(outline)) {
-         clean$outline <- FALSE
-         return(clean)
-      }
-      clean$outline <- TRUE
-      # setting: caption
-      if(is.null(outline$caption)) clean$caption <- defaultCaption
-      else {
-         if(!("character" %in% class(outline$caption))) stop("PivotDataGroup$cleanOutlineArg(): The caption for the outline data group must be a character value.", call. = FALSE)
-         clean$caption <- outline$caption
-      }
-      # setting: isEmpty
-      if(is.null(outline$isEmpty)) clean$isEmpty <- defaultIsEmpty
-      else {
-         if(!("logical" %in% class(outline$isEmpty))) stop("PivotDataGroup$cleanOutlineArg(): The isEmpty value for the outline data group must be a logical (TRUE/FALSE) value.", call. = FALSE)
-         clean$isEmpty <- outline$isEmpty
-      }
-      # setting: mergeSpace
-      allowedMergeSpaceValues <- c("doNotMerge", "dataGroupsOnly", "cellsOnly", "dataGroupsAndCellsAs1", "dataGroupsAndCellsAs2")
-      if(is.null(outline$mergeSpace)) clean$mergeSpace <- ifelse(defaultIsEmpty, "dataGroupsAndCellsAs2", "dataGroupsOnly")
-      else {
-         if(!("character" %in% class(outline$mergeSpace))) stop("PivotDataGroup$cleanOutlineArg(): The mergeSpace value for the outline data group must be a character value.", call. = FALSE)
-         if(!(outline$mergeSpace %in% allowedMergeSpaceValues)) stop("PivotDataGroup$cleanOutlineArg(): The mergeSpace value for the outline data group must be one of the following values: doNotMerge, dataGroupsOnly, cellsOnly, dataGroupsAndCellsAs1, dataGroupsAndCellsAs2.", call. = FALSE)
-         clean$mergeSpace <- outline$mergeSpace
-      }
-      # setting: styling
-      allowedStylingValues <- c("plain", "outline")
-      if(is.null(outline$styling)) clean$styling <- "outline"
-      else {
-         if(!("character" %in% class(outline$styling))) stop("PivotDataGroup$cleanOutlineArg(): The styling value for the outline data group must be a character value.", call. = FALSE)
-         if(!(outline$styling %in% allowedStylingValues)) stop("PivotDataGroup$cleanOutlineArg(): The styling value for the outline data group must be one of the following values: plain, outline.", call. = FALSE)
-         clean$styling <- outline$styling
-      }
-      # setting: groupStyleName (allow null)
-      if(!is.null(outline$groupStyleName)) {
-         if(!("character" %in% class(outline$groupStyleName))) stop("PivotDataGroup$cleanOutlineArg(): The groupStyleName for the outline data group must be a character value.", call. = FALSE)
-         clean$groupStyleName <- outline$groupStyleName
-      }
-      # setting: groupStyleDeclarations (allow null)
-      if(!is.null(outline$groupStyleDeclarations)) {
-         if(!("list" %in% class(outline$groupStyleDeclarations))) stop("PivotDataGroup$cleanOutlineArg(): The groupStyleDeclarations for the outline data group must a list.", call. = FALSE)
-         clean$groupStyleDeclarations <- outline$groupStyleDeclarations
-      }
-      # setting: cellStyleName (allow null)
-      if(!is.null(outline$cellStyleName)) {
-         if(!("character" %in% class(outline$cellStyleName))) stop("PivotDataGroup$cleanOutlineArg(): The cellStyleName for the outline data group must be a character value.", call. = FALSE)
-         clean$cellStyleName <- outline$cellStyleName
-      }
-      # setting: cellStyleDeclarations (allow null)
-      if(!is.null(outline$cellStyleDeclarations)) {
-         if(!("list" %in% class(outline$cellStyleDeclarations))) stop("PivotDataGroup$cleanOutlineArg(): The cellStyleDeclarations for the outline data group must a list.", call. = FALSE)
-         clean$cellStyleDeclarations <- outline$cellStyleDeclarations
-      }
-      # check mergeSpace and isEmpty are compatible
-      if(!(clean$isEmpty)) {
-         if(clean$mergeSpace=="cellsOnly") clean$mergeSpace <- "doNotMerge"
-         else if(clean$mergeSpace %in% c("dataGroupsAndCellsAs1", "dataGroupsAndCellsAs2")) clean$mergeSpace <- "dataGroupsOnly"
-      }
-      if(private$p_parentPivot$traceEnabled==TRUE) private$p_parentPivot$trace("PivotDataGroup$cleanOutlineArg", "Cleaned outline argument.")
-      return(clean)
-   },
    createOutlineGroup = function(parentGroup=NULL, outline=NULL, variableName=NULL, values=NULL, caption=NULL,
                                  calculationGroupName=NULL, calculationName=NULL,
                                  baseStyleName=NULL, styleDeclarations=NULL, isLevelTotal=FALSE,