+++ R/utilities.R
@@ -160,3 +160,90 @@ exportValueAs <- function(rawValue, formattedValue, exportOptions, blankValue=ch
   return(invisible(formattedValue))
 }
 
+#' Clean the arguments specified for an outline group
+#'
+#' \code{cleanOutlineArg} checks values and provides defaults.
+#'
+#' @param pvt The pivot table.
+#' @param outline Either a logical value (TRUE to use the default
+#' outline settings) or a list specifying outline settings.
+#' @param defaultCaption The default caption of the outline group.
+#' @param defaultIsEmpty Specify whether the outline group is
+#' empty or contains a value (typically a sub-total)
+#' @return A listed containing checked/cleaned outline group
+#' settings.
+
+cleanOutlineArg = function(pvt, outline=NULL, defaultCaption="{value}", defaultIsEmpty=TRUE) { # checks values and provides defaults for settings (so following code does not need is.null checks)
+  if(pvt$argumentCheckMode > 0) {
+    checkArgument(pvt$argumentCheckMode, FALSE, "(global)", "cleanOutlineArg", outline, missing(outline), allowMissing=FALSE, allowNull=TRUE, allowedClasses=c("logical", "list"))
+    checkArgument(pvt$argumentCheckMode, FALSE, "(global)", "cleanOutlineArg", defaultCaption, missing(defaultCaption), allowMissing=TRUE, allowNull=FALSE, allowedClasses="character")
+    checkArgument(pvt$argumentCheckMode, FALSE, "(global)", "cleanOutlineArg", defaultIsEmpty, missing(defaultIsEmpty), allowMissing=TRUE, allowNull=FALSE, allowedClasses="logical")
+  }
+  if(pvt$traceEnabled==TRUE) pvt$trace("PivotDataGroup$cleanOutlineArg", "Cleaning outline argument...")
+  if("logical" %in% class(outline)) {
+    if(isTRUE(outline)) outline <- list()
+    else outline <- NULL
+  }
+  clean <- list()
+  # global switch
+  if(is.null(outline)) {
+    clean$outline <- FALSE
+    return(clean)
+  }
+  clean$outline <- TRUE
+  # setting: caption
+  if(is.null(outline$caption)) clean$caption <- defaultCaption
+  else {
+    if(!("character" %in% class(outline$caption))) stop("cleanOutlineArg(): The caption for the outline data group must be a character value.", call. = FALSE)
+    clean$caption <- outline$caption
+  }
+  # setting: isEmpty
+  if(is.null(outline$isEmpty)) clean$isEmpty <- defaultIsEmpty
+  else {
+    if(!("logical" %in% class(outline$isEmpty))) stop("cleanOutlineArg(): The isEmpty value for the outline data group must be a logical (TRUE/FALSE) value.", call. = FALSE)
+    clean$isEmpty <- outline$isEmpty
+  }
+  # setting: mergeSpace
+  allowedMergeSpaceValues <- c("doNotMerge", "dataGroupsOnly", "cellsOnly", "dataGroupsAndCellsAs1", "dataGroupsAndCellsAs2")
+  if(is.null(outline$mergeSpace)) clean$mergeSpace <- ifelse(defaultIsEmpty, "dataGroupsAndCellsAs2", "dataGroupsOnly")
+  else {
+    if(!("character" %in% class(outline$mergeSpace))) stop("cleanOutlineArg(): The mergeSpace value for the outline data group must be a character value.", call. = FALSE)
+    if(!(outline$mergeSpace %in% allowedMergeSpaceValues)) stop("cleanOutlineArg(): The mergeSpace value for the outline data group must be one of the following values: doNotMerge, dataGroupsOnly, cellsOnly, dataGroupsAndCellsAs1, dataGroupsAndCellsAs2.", call. = FALSE)
+    clean$mergeSpace <- outline$mergeSpace
+  }
+  # setting: styling
+  allowedStylingValues <- c("plain", "outline")
+  if(is.null(outline$styling)) clean$styling <- "outline"
+  else {
+    if(!("character" %in% class(outline$styling))) stop("cleanOutlineArg(): The styling value for the outline data group must be a character value.", call. = FALSE)
+    if(!(outline$styling %in% allowedStylingValues)) stop("cleanOutlineArg(): The styling value for the outline data group must be one of the following values: plain, outline.", call. = FALSE)
+    clean$styling <- outline$styling
+  }
+  # setting: groupStyleName (allow null)
+  if(!is.null(outline$groupStyleName)) {
+    if(!("character" %in% class(outline$groupStyleName))) stop("cleanOutlineArg(): The groupStyleName for the outline data group must be a character value.", call. = FALSE)
+    clean$groupStyleName <- outline$groupStyleName
+  }
+  # setting: groupStyleDeclarations (allow null)
+  if(!is.null(outline$groupStyleDeclarations)) {
+    if(!("list" %in% class(outline$groupStyleDeclarations))) stop("cleanOutlineArg(): The groupStyleDeclarations for the outline data group must a list.", call. = FALSE)
+    clean$groupStyleDeclarations <- outline$groupStyleDeclarations
+  }
+  # setting: cellStyleName (allow null)
+  if(!is.null(outline$cellStyleName)) {
+    if(!("character" %in% class(outline$cellStyleName))) stop("cleanOutlineArg(): The cellStyleName for the outline data group must be a character value.", call. = FALSE)
+    clean$cellStyleName <- outline$cellStyleName
+  }
+  # setting: cellStyleDeclarations (allow null)
+  if(!is.null(outline$cellStyleDeclarations)) {
+    if(!("list" %in% class(outline$cellStyleDeclarations))) stop("cleanOutlineArg(): The cellStyleDeclarations for the outline data group must a list.", call. = FALSE)
+    clean$cellStyleDeclarations <- outline$cellStyleDeclarations
+  }
+  # check mergeSpace and isEmpty are compatible
+  if(!(clean$isEmpty)) {
+    if(clean$mergeSpace=="cellsOnly") clean$mergeSpace <- "doNotMerge"
+    else if(clean$mergeSpace %in% c("dataGroupsAndCellsAs1", "dataGroupsAndCellsAs2")) clean$mergeSpace <- "dataGroupsOnly"
+  }
+  if(pvt$traceEnabled==TRUE) pvt$trace("cleanOutlineArg", "Cleaned outline argument.")
+  return(clean)
+}